refactor: Make CAPI data optional when building page html
